# perf: bypass macOS fs read system cache to avoid multi-threads lock