--- a/src/lib.rs
+++ b/src/lib.rs
@@ -52,6 +52,10 @@ mod cache;
 mod context;
 mod error;
 mod file_system;
+#[cfg(target_os = "macos")]
+mod macos;
+#[cfg(target_os = "macos")]
+mod macos_cache;
 mod options;
 mod package_json;
 mod path;
@@ -95,6 +99,10 @@ pub use crate::{
         CompilerOptions, CompilerOptionsPathsMap, ExtendsField, ProjectReference, TsConfig,
     },
 };
+
+#[cfg(target_os = "macos")]
+pub use crate::{macos::MacOsFs, macos_cache::PackageJsonCache};
+
 use crate::{
     context::ResolveContext as Ctx, path::SLASH_START, specifier::Specifier,
     tsconfig_context::TsconfigResolveContext,