--- a/src/lib.rs
+++ b/src/lib.rs
@@ -54,8 +54,6 @@ mod error;
 mod file_system;
 #[cfg(target_os = "macos")]
 mod macos;
-#[cfg(target_os = "macos")]
-mod macos_cache;
 mod options;
 mod package_json;
 mod path;
@@ -101,7 +99,7 @@ pub use crate::{
 };
 
 #[cfg(target_os = "macos")]
-pub use crate::{macos::MacOsFs, macos_cache::PackageJsonCache};
+pub use crate::macos::MacOsFs;
 
 use crate::{
     context::ResolveContext as Ctx, path::SLASH_START, specifier::Specifier,