Introduce a macOS-only fast path in `MacOsFs::read_nocache` that opens files with `open` and immediately toggles `F_NOCACHE` via `fcntl`. Skipping the unified buffer cache keeps resolver threads from fighting over the kernel's shared cache lock, which yields better latency under multi-threaded lookups.

<details>
<summary>
The performance improvement observed was due to `PackageJsonCache`, not the `F_NOCACHE` optimization.
</summary>

```
resolver/single-thread  time:   [70.079 µs 71.033 µs 72.051 µs]
                        change: [−7.3466% −4.1720% −0.6042%] (p = 0.01 < 0.05)
                        Change within noise threshold.
Found 3 outliers among 100 measurements (3.00%)
  1 (1.00%) high mild
  2 (2.00%) high severe
resolver/multi-thread   time:   [58.623 µs 59.441 µs 60.257 µs]
                        change: [−12.697% −7.1853% −2.2826%] (p = 0.01 < 0.05)
                        Performance has improved.
Found 4 outliers among 100 measurements (4.00%)
  3 (3.00%) high mild
  1 (1.00%) high severe
resolver/resolve from symlinks
                        time:   [24.349 ms 24.580 ms 24.816 ms]
                        change: [−19.254% −18.294% −17.360%] (p = 0.00 < 0.05)
                        Performance has improved.
```

</details>

Co-authored-by: Codex <codex@openai.com>