@Boshen 2025-09-18
We should probably do this in stages, first PR don't need PackageJsonCache.