@Brooooooklyn 2025-09-18
@Boshen The `PackageJsonCache` implementation was removed, and then no performance changes in the benchmark. It may be necessary to test on a larger project with real-world workloads to verify the impact.