@Brooooooklyn 2025-09-18
The results of the benchmark may be very inaccurate because the file read results from previous runs may be cached by the system during repeated executions.